feat(终端): 像素级滚动——分数视口+终端设置开关（旧行级保底）
设计契约 docs/active/设置页.md §三（2026-09-24 用户拍板「滚动的像素级」）：
- scroll.rs moved_px：同 slop 门不取整原样出位移，越阈第一笔 slop 段不计入
- termview scroll_px 分数视口：零头累计满行提交 display_offset，两端钳制
  （贴底不许负零头/贴历史顶不许正零头），拨关清亚行态；渲染双路径
  （softbuffer/GPU）正零头顶缘多收 line=-1、负零头底缘多收 line=screen_lines
- GpuCell.py i64 化（多收行原点可负）；GLES 合成期实例平移
  dy - kb_frac + scroll_frac，present_frame scissor 带 (顶带内缘, 内容底沿)
- 触摸分流 pixel_lane（一次触摸内不换道；tmux 全屏鼠标上报照旧 SGR 滚轮）
  + 启动读回 terminal.json pixelScroll
- 设置页下池第二行「终端设置/滚动行为」→ 上池「像素级滚动」下拉
  （关=行级保底·默认/开=像素级实验），写盘+即时分流
考题 scroll_spec +2 / termview_spec +3 / settings_spec +1（含关时渲染
零差异钉——旧保底开关关=零行为）；变异三咬全中：多收行删除→红、
贴底钳制删除→红、moved_px 取整回潮→红，复原全绿
另含 BAR-152 bugs.md 用户终验结案行（前批 3b87a73 的账本收尾）

+++ src/glyph_atlas.rs
@@ -200,7 +200,10 @@ impl GlyphAtlas {
 #[derive(Clone, Copy, Debug)]
 pub struct GpuCell {
     pub px: u32,
-    pub py: u32,
+    /// 有符号格原点（2026-09-24 像素级滚动）：顶缘多收的历史行
+    /// py = margin_top - cell_h 为负——零头归合成期实例平移，负原点
+    /// 经 dy 拉回顶带内，越顶带部分由 grid_clip scissor 裁
+    pub py: i64,
     pub fg: u32,
     pub bg: u32,
     pub c: char,
@@ -305,7 +308,7 @@ pub fn grid_to_instances(
         }
         out.glyph.push(GlyphInstance {
             x: left as f32,
-            y: (cell.py as i64 + i64::from(slot.off_y)) as f32,
+            y: (cell.py + i64::from(slot.off_y)) as f32,
             w: draw_w as f32,
             h: f32::from(slot.h),
             u0: f32::from(slot.u0) / page.w as f32,
